refactor: 🚚 moved datathon constants to `/utils` folder
