+++ src/db/utils.ts
@@ -1,9 +1,10 @@
-import { isDatathonWeek } from '@/app/datathon/data';
 import connectDB from '@/db/config';
 import LeaderboardModel, {
   type LeaderboardContent,
   type LeaderboardType,
 } from '@/db/models/leaderboard';
+
+import { isDatathonWeek } from '@/utils/datathon';
 
 export const getLeaderboardData = async (
   type: LeaderboardType = 'public',
